Add first Bash spec
A simple file with all in. I will keep it as it is for now for better reuse, where we can easily download with a single line. 

What can we improve? We can split it and use a ci to generates the big file, as we see in some js libraries, including a minification of the content.